implement login command
