+++ cmd/util.go
@@ -1,0 +1,21 @@
+package cmd
+
+import (
+	"os"
+	"os/signal"
+)
+
+func trapSignal(cleanupFunc func()) {
+	c := make(chan os.Signal, 1)
+	signal.Notify(c, os.Interrupt)
+
+	go func() {
+		// block until we've received a signal
+		<-c
+
+		// execute any cleanup logic
+		if cleanupFunc != nil {
+			cleanupFunc()
+		}
+	}()
+}
