+++ go.mod
@@ -3,13 +3,14 @@ module github.com/cosmos/atlas
 go 1.15
 
 require (
+	github.com/BurntSushi/toml v0.3.1
 	github.com/dghubble/gologin/v2 v2.2.0
-	github.com/go-playground/validator/v10 v10.3.0
-	github.com/golang-migrate/migrate/v4 v4.12.2
+	github.com/fatih/color v1.9.0
+	github.com/go-playground/validator/v10 v10.4.0
+	github.com/golang-migrate/migrate/v4 v4.13.0
 	github.com/google/go-github v17.0.0+incompatible
 	github.com/gorilla/mux v1.8.0
 	github.com/gorilla/sessions v1.2.1
-	github.com/jackc/pgproto3/v2 v2.0.4 // indirect
 	github.com/joho/godotenv v1.3.0
 	github.com/justinas/alice v1.2.0
 	github.com/knadh/koanf v0.13.0
@@ -18,8 +19,8 @@ require (
 	github.com/satori/go.uuid v1.2.0
 	github.com/stretchr/testify v1.6.1
 	github.com/urfave/cli/v2 v2.2.0
-	golang.org/x/crypto v0.0.0-20200820211705-5c72a883971a // indirect
+	golang.org/x/crypto v0.0.0-20201002170205-7f63de1d35b0 // indirect
 	golang.org/x/oauth2 v0.0.0-20200902213428-5d25da1a8d43
-	gorm.io/driver/postgres v1.0.1
-	gorm.io/gorm v1.20.1
+	gorm.io/driver/postgres v1.0.2
+	gorm.io/gorm v1.20.2
 )
